requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.6.0
astunparse==1.6.3
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
fastapi==0.115.0
ffmpy==0.4.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.9.0
gast==0.6.0
google-pasta==0.2.0
gradio==4.44.1
gradio_client==1.3.0
grpcio==1.66.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.8
importlib_resources==6.4.5
Jinja2==3.1.4
keras==3.5.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.0
namex==0.0.8
numpy==1.26.4
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pillow==10.4.0
protobuf==4.25.4
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.8.1
ruff==0.6.8
semantic-version==2.10.0
setuptools==74.1.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow @ file:///C:/Users/HP/PROJECT%20GROUP/ISIC%20IMAGE/tensorflow-2.17.0-cp312-cp312-win_amd64.whl#sha256=278bc80642d799adf08dc4e04f291aab603bba7457d50c1f9bc191ebbca83f43
tensorflow-intel==2.17.0
termcolor==2.4.0
tomlkit==0.12.0
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.0
websockets==12.0
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
